+++ includes/class-loader.php
@@ -31,6 +31,7 @@ $alynt_ag_files = array(
 	'includes/services/class-frontend-assets.php',
 	'includes/services/class-frontend-branding.php',
 	'includes/services/class-frontend-components.php',
+	'includes/services/class-offcanvas-menu-walker.php',
 	'includes/services/class-frontend-dashboard-screen.php',
 	'includes/services/class-frontend-register-screen.php',
 	'includes/services/class-frontend-login-screen.php',